MA-15988[Android] Use Quickstep as default launcher on normal android
As enabled Multi-display feature, there is a popup menu to select the launcher.
Here remove the DEFAULT category of Multi-display launcher to make the
Quickstep as default launcher and no need to select launcher when boot up.
If need to evaluate the Multi-display feature, user can open the "MD Launcher"
app to get the multi-display launcher.

Change-Id: I6a119c0e31a0eb954f9d2386419247963c801e60
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>